Add StoredFile and StoredList table to SQLite database

+++ project/src/com/lasthopesoftware/bluewater/data/sqlite/access/DatabaseHandler.java
@@ -15,23 +15,34 @@ import com.j256.ormlite.support.ConnectionSource;
 import com.j256.ormlite.table.DatabaseTableConfig;
 import com.j256.ormlite.table.TableUtils;
 import com.lasthopesoftware.bluewater.data.sqlite.objects.Library;
+import com.lasthopesoftware.bluewater.data.sqlite.objects.StoredFile;
+import com.lasthopesoftware.bluewater.data.sqlite.objects.StoredList;
 
 public class DatabaseHandler extends OrmLiteSqliteOpenHelper  {
 
-	private static int DATABASE_VERSION = 2;
+	private static int DATABASE_VERSION = 3;
 	private static final String DATABASE_NAME = "sessions_db";
 	
 	@SuppressWarnings("rawtypes")
-	private static Class[] tables = { Library.class };
+	private static Class[] version2Tables = { Library.class };
+	@SuppressWarnings("rawtypes")
+	private static Class[] version3Tables = { StoredFile.class, StoredList.class };
+	@SuppressWarnings("rawtypes")
+	private static Class[][] allTables = { version2Tables, version3Tables };
 	
 	public DatabaseHandler(Context context) {
 		super(context, DATABASE_NAME, null, DATABASE_VERSION);
 	}
 
-	@SuppressWarnings({ "unchecked", "rawtypes" })
+	@SuppressWarnings("rawtypes")
 	@Override
 	public void onCreate(SQLiteDatabase db, ConnectionSource conn) {
-		for (Class table : tables) {
+		for (Class[] tableArray : allTables) createTables(conn, tableArray);
+	}
+	
+	@SuppressWarnings({ "unchecked", "rawtypes" })
+	private void createTables(ConnectionSource conn, Class[] tableClasses) {
+		for (Class table : tableClasses) {
 			try {
 				TableUtils.createTable(conn, table);
 			} catch (SQLException e) {
@@ -42,15 +53,20 @@ public class DatabaseHandler extends OrmLiteSqliteOpenHelper  {
 
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	@Override
-	public void onUpgrade(SQLiteDatabase db, ConnectionSource conn, int arg2, int arg3) {
-		for (Class table : tables) {
-			try {
-				TableUtils.dropTable(conn, table, true);
-			} catch (SQLException e) {
-				LoggerFactory.getLogger(DatabaseHandler.class).error(e.toString(), e);
+	public void onUpgrade(SQLiteDatabase db, ConnectionSource conn, int oldVersion, int newVersion) {
+		if (oldVersion < 2) {
+			for (Class table : version2Tables) {
+				try {
+					TableUtils.dropTable(conn, table, true);
+				} catch (SQLException e) {
+					LoggerFactory.getLogger(DatabaseHandler.class).error(e.toString(), e);
+				}
 			}
+			createTables(conn, version2Tables);
 		}
-		onCreate(db, conn);
+		
+		if (oldVersion < 3)
+			createTables(conn, version3Tables);
 	}
 	
 	public <D extends Dao<T, ?>, T> D getAccessObject(Class<T> c) throws SQLException  {
